2 games including game logic and content
THIS IS STAGE1:
this are unfinished versions of both games, I left them in this stage, because we will need to debate and decide which approach to take. Both games have multiple ways to integrate and implement solutions.

STAGE2:
Will be to implement correctly the game mods and controllers which Jack has provided (basically do duplicates of his C++ code or do Blueprint versions of it, also taylor these minigames for our needs.
EXAMPLE: SolarCars mini games is for now build only for one player, but it is very easy to rebuild this game for 4 players, enlarge the "track" or change obstacles for clouds (dark paths you should avoid)
SolarCars game is very modular and easy to edit.

STAGE3:
Finishing the games
After we will decide what is doable we can start to implement the rest of the game logic and polish it.

IT IS A MESS
I did as much as I could with all the naming conventions etc BUT it is still messy and there is a lot of EXPERIMENTAL stuff left over, in case that we will decide to use them or not.