Major commit:
1) Added two mini games including assets

Conveyor mini game is WIP but nearly finished, everything should be set up correctly and if controllers are available then test play should be done.

TO DO:

Spawners are spawning "random" sized cubes as default objects, this might be changed.
Casey kindly provided me a few assets, but these need to be resized and blueprints created out of them must be done, also player shouldnt just walk over these obstacles, he should be able to jump over them.
On the top of that, timer delay is 2 seconds hardcoded solution. There might be added random values.

There is a lot of stuff in the Experimental folder such as objects, spawners and everything else I have used durring the development and hadn't used for any reason.

SOLAR CARS GAME:

Map is procedurally generated and at this moment only for reference is spawning way too big middle floor (only for Elliot), i have also realised that the middle curved floor could be scaled only on X axis not on Y axis, but everything is modular so there shouldn't be any problem to change it / play with it

The biggest question is player character, I am for moveable cars, which can turn left and right to avoid cloud shadows (I will talk about them in a second) and use jumps (again in experimental folder hopefully)

for now there is a CLOUD NINE spawner which is spawning clouds in a row, again it is modular and can values are easy to change.

About the clouds then: Decals are working, but there is a issue connected to them. All lighting has turned off the "shadow casting" function, so scene looks overall bright and "sunny" once I spawn a cloud, the dark patch is not really well visible, perhaps I (WE) can do some particle effect from the cloud so it will be properly visible, next added feature are volume triggers which are in the exact shape as decals which are the exact shape of a cloud and are scalable. this will need some fine tuning and fidling around.

Another Issue connected with this game is the score system, where we will need to discuss it on the next meeting.

My commit is a bit messy with the items and naming conventions, I'll fix that until the end of week, this commit is done so Elliot can now fork, clone and have a look arround.

I am here to answer any questions concerning these two mini games.

So this one is having multiple problems to be solved